update machines and alter hash(object) to hashlb.md5

+++ build.py
@@ -4,6 +4,7 @@ We are going to build machines in machines.csv.
 """
 import os
 import sys
+import hashlib
 from pyquery import PyQuery as pq
 
 
@@ -24,12 +25,8 @@ def find_urls_in_openwrt_homepage(homepage):
             dot_config = os.path.join(homepage, href)
     return image_builder, dot_config
 
-
-def build():
+def build_hash_of_url_to_image_builder():
     header = None
-    # we found a .config always in an ImageBuilder
-    # we also found many firmware share the same ImageBuilder
-    # because they are downloaded from the same page
     image_builder_table = {}
     with open('machines.csv') as f:
         for line in f:
@@ -41,16 +38,25 @@ def build():
             url = items[header.index('url')]
             homepage = os.path.dirname(url)
             url_to_image_builder, _ = find_urls_in_openwrt_homepage(homepage)
-            hash_of_image_builder = hash(url_to_image_builder)
-            if hash_of_image_builder in image_builder_table:
-                image_builder_table[hash(url_to_image_builder)]['support'].append(uuid)
+            hash_of_image_builder = hashlib.md5(url_to_image_builder.encode('utf-8'))
+            if hash_of_image_builder.hexdigest() in image_builder_table:
+                # we also found many firmware share the same ImageBuilder
+                # because they are downloaded from the same page
+                image_builder_table[hash_of_image_builder.hexdigest()]['support'].append(uuid)
             else:
-                image_builder_table[hash(url_to_image_builder)] = {'url': url_to_image_builder, 'support': [uuid]}
-    # save image_builder_table
-    with open('image_builder.csv', 'w') as f:
-        for k, v in image_builder_table.items():
-            f.write('{},{},{}\n'.format(k, v['url'], ','.join(v['support'])))
-    # download the image_builder to share
+                image_builder_table[hash_of_image_builder.hexdigest()] = {'url': url_to_image_builder, 'support': [uuid]}
+    return image_builder_table
+
+
+def build():
+    # we found a .config always in an ImageBuilder
+    if not os.path.exists('image_builder.csv'):
+        image_builder_table = build_hash_of_url_to_image_builder()
+        # save image_builder_table
+        with open('image_builder.csv', 'w') as f:
+            for k, v in image_builder_table.items():
+                f.write('{},{},{}\n'.format(k, v['url'], ','.join(v['support'])))
+    # download the image_builder to ./share
     for k, v in image_builder_table.items():
         os.system('wget -nc {} -P share'.format(v['url']))
         # TODO
